backend_ats/requirements.txt
bcrypt==4.2.0
python-dotenv==1.0.0
uuid==1.30
sqlalchemy==2.0.27
PyJWT==2.8.0
Flask==2.3.3
Flask-Cors==4.0.0
openai==1.14.0
PyPDF2==3.0.1
tiktoken==0.5.2
werkzeug
qdrant-client==1.7.0
httpx==0.25.0
psycopg2-binary==2.9.9
matplotlib==3.10.7